+++ spec2nii/Philips/philips_dcm.py
@@ -2,12 +2,16 @@
 Author: William Clarke <william.clarke@ndcn.ox.ac.uk>
 Copyright (C) 2020 University of Oxford
 """
+from datetime import datetime
+
 import numpy as np
 import nibabel.nicom.dicomwrappers
+
+from nifti_mrs.create_nmrs import gen_nifti_mrs_hdr_ext
+from nifti_mrs.hdr_ext import Hdr_Ext
+
 from spec2nii.dcm2niiOrientation.orientationFuncs import dcm_to_nifti_orientation
 from spec2nii.nifti_orientation import NIFTIOrient
-from spec2nii import nifti_mrs
-from datetime import datetime
 from spec2nii import __version__ as spec2nii_ver
 
 
@@ -155,9 +159,9 @@ def multi_file_dicom(files_in, fname_out, tag, verbose):
             meta_used.set_dim_info(0, tag)
 
         # Create NIFTI MRS object.
-        nifti_mrs_out.append(nifti_mrs.NIfTI_MRS(combined_data, or_used.Q44, dt_used, meta_used))
+        nifti_mrs_out.append(gen_nifti_mrs_hdr_ext(combined_data, dt_used, meta_used, or_used.Q44))
         if ref_list[0] is not None:
-            nifti_mrs_out.append(nifti_mrs.NIfTI_MRS(combined_ref, or_used.Q44, dt_used, meta_ref_used))
+            nifti_mrs_out.append(gen_nifti_mrs_hdr_ext(combined_ref, dt_used, meta_ref_used, or_used.Q44))
     else:
         for idx, (dd, rr, oo, dt, mm, mr, ff) in enumerate(zip(data_list,
                                                                ref_list,
@@ -169,12 +173,12 @@ def multi_file_dicom(files_in, fname_out, tag, verbose):
             # Add original files to nifti meta information.
             mm.set_standard_def('OriginalFile', [str(ff.name), ])
             fnames_out.append(f'{mainStr}_{idx:03}')
-            nifti_mrs_out.append(nifti_mrs.NIfTI_MRS(dd, oo.Q44, dt, mm))
+            nifti_mrs_out.append(gen_nifti_mrs_hdr_ext(dd, dt, mm, oo.Q44))
 
             if rr is not None:
                 mr.set_standard_def('OriginalFile', [str(ff.name), ])
                 fnames_out.append(f'{mainStr}_ref_{idx:03}')
-                nifti_mrs_out.append(nifti_mrs.NIfTI_MRS(rr, oo.Q44, dt, mr))
+                nifti_mrs_out.append(gen_nifti_mrs_hdr_ext(rr, dt, mr, oo.Q44))
 
     return nifti_mrs_out, fnames_out
 
@@ -357,8 +361,9 @@ def _extractDicomMetadata_old(dcmdata, water_suppressed=True):
     """
 
     # Extract required metadata and create hdr_ext object
-    obj = nifti_mrs.hdr_ext(dcmdata.dcm_data.TransmitterFrequency,
-                            dcmdata.dcm_data.ResonantNucleus)
+    obj = Hdr_Ext(
+        dcmdata.dcm_data.TransmitterFrequency,
+        dcmdata.dcm_data.ResonantNucleus)
 
     # private_mrs_tags = dcmdata[('2005', '140f')][0]
 
@@ -461,8 +466,9 @@ def _extractDicomMetadata_new(dcmdata, water_suppressed=True):
     """
 
     # Extract required metadata and create hdr_ext object
-    obj = nifti_mrs.hdr_ext(dcmdata.dcm_data.TransmitterFrequency,
-                            dcmdata.dcm_data.ResonantNucleus)
+    obj = Hdr_Ext(
+        dcmdata.dcm_data.TransmitterFrequency,
+        dcmdata.dcm_data.ResonantNucleus)
 
     def set_if_present(tag, val):
         if val: