+++ spec2nii/Siemens/twixfunctions.py
@@ -7,10 +7,11 @@ from os.path import basename
 from datetime import datetime
 
 import numpy as np
+from nifti_mrs.create_nmrs import gen_nifti_mrs_hdr_ext
+from nifti_mrs.hdr_ext import Hdr_Ext
 
 import spec2nii.GSL.gslfunctions as GSL
 from spec2nii.dcm2niiOrientation.orientationFuncs import dcm_to_nifti_orientation
-from spec2nii import nifti_mrs
 from spec2nii import __version__ as spec2nii_ver
 
 
@@ -527,7 +528,7 @@ def assemble_nifti_mrs(data, dwellTime, orientation, meta_obj, dim_tags=None):
         for idx, dt in zip(range(data.ndim - 4), dim_tags):
             meta_obj.set_dim_info(idx, dt)
 
-    return nifti_mrs.NIfTI_MRS(data, orientation.Q44, dwellTime, meta_obj)
+    return gen_nifti_mrs_hdr_ext(data, dwellTime, meta_obj, orientation.Q44)
 
 
 def twix2DCMOrientation(mapVBVDHdr, force_svs=False, verbose=False):
@@ -690,8 +691,9 @@ def extractTwixMetadata_xa(mapVBVDHdr, original_file):
     """
 
     # Extract required metadata and create hdr_ext object
-    obj = nifti_mrs.hdr_ext(mapVBVDHdr['Dicom'][('lFrequency')] / 1E6,
-                            mapVBVDHdr['MeasYaps'][('sTXSPEC', 'asNucleusInfo', '0', 'tNucleus')].strip('"'))
+    obj = Hdr_Ext(
+        mapVBVDHdr['Dicom'][('lFrequency')] / 1E6,
+        mapVBVDHdr['MeasYaps'][('sTXSPEC', 'asNucleusInfo', '0', 'tNucleus')].strip('"'))
 
     # Standard defined metadata
     # # 5.1 MRS specific Tags
@@ -802,8 +804,9 @@ def extractTwixMetadata_vx(mapVBVDHdr, original_file):
     """
 
     # Extract required metadata and create hdr_ext object
-    obj = nifti_mrs.hdr_ext(mapVBVDHdr['Meas'][('Frequency')] / 1E6,
-                            mapVBVDHdr['Meas'][('ResonantNucleus')])
+    obj = Hdr_Ext(
+        mapVBVDHdr['Meas'][('Frequency')] / 1E6,
+        mapVBVDHdr['Meas'][('ResonantNucleus')])
 
     # Standard defined metadata
     # # 5.1 MRS specific Tags