+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This document contains the Spec2nii release history in reverse chronological ord
 - Added special-case handling of the mgs_svs_ed(_universal) for twix data inputs.
 - Modified repository structure. Siemens and Philips files now have dedicated directories.
 - Support Python 3.11
+- Enable automatic Pypi upload and move to pyproject.toml file build.
 
 0.4.9 (Friday 4th November 2022)
 --------------------------------